Moved Radio Button component to ui directory
